Minor reorganization of osqueryi linking
