+++ osquery/CMakeLists.txt
@@ -47,9 +47,6 @@ else()
   ADD_OSQUERY_LINK_CORE("libz")
   ADD_OSQUERY_LINK_CORE("boost_system-mt")
   ADD_OSQUERY_LINK_CORE("boost_filesystem-mt")
-  ADD_OSQUERY_LINK_CORE("boost_regex-mt")
-  ADD_OSQUERY_LINK_CORE("rocksdb_lite")
-  ADD_OSQUERY_LINK_CORE("snappy")
   ADD_OSQUERY_LINK_CORE("gflags")
   ADD_OSQUERY_LINK_CORE("thrift")
   ADD_OSQUERY_LINK_CORE("lz4")
@@ -72,11 +69,16 @@ ADD_OSQUERY_LINK_ADDITIONAL("cppnetlib-uri")
 ADD_OSQUERY_LINK_ADDITIONAL("cppnetlib-client-connections")
 
 if(NOT WINDOWS)
+  ADD_OSQUERY_LINK_ADDITIONAL("boost_regex-mt")
+  ADD_OSQUERY_LINK_ADDITIONAL("rocksdb_lite")
+  ADD_OSQUERY_LINK_ADDITIONAL("snappy")
   ADD_OSQUERY_LINK_ADDITIONAL("ssl")
   ADD_OSQUERY_LINK_ADDITIONAL("yara")
-  ADD_OSQUERY_LINK_ADDITIONAL("linenoise")
   ADD_OSQUERY_LINK_ADDITIONAL("crypto")
   ADD_OSQUERY_LINK_CORE("crypto")
+
+  # Linenoise will be used when compiling osqueryi.
+  ADD_OSQUERY_LINK_ADDITIONAL("linenoise")
 endif()
 
 if(APPLE)
@@ -104,19 +106,6 @@ elseif(LINUX OR FREEBSD)
   if(NOT DEFINED ENV{SANITIZE} AND NOT DEFINED ENV{DEBUG})
     ADD_OSQUERY_LINK_CORE("-pie")
   endif()
-endif()
-
-if(LINUX)
-  ADD_OSQUERY_LINK_CORE("libgcc_s")
-  if(DEFINED ENV{OSQUERY_BUILD_SHARED})
-    ADD_OSQUERY_LINK_ADDITIONAL("-L${BUILD_DEPS}/legacy/lib")
-  endif()
-  if(NOT DEFINED ENV{OSQUERY_BUILD_LINK_SHARED})
-    ADD_OSQUERY_LINK_CORE("-static-libstdc++")
-    ADD_OSQUERY_LINK_ADDITIONAL("-static-libstdc++")
-  endif()
-  # For Ubuntu/CentOS packages add the build SHA1.
-  ADD_OSQUERY_LINK_CORE("-Wl,--build-id")
 endif()
 
 if(DEFINED ENV{SANITIZE})
@@ -154,6 +143,20 @@ if(NOT DEFINED ENV{SKIP_TABLES})
   GENERATE_UTILITIES("${CMAKE_SOURCE_DIR}")
   AMALGAMATE("${CMAKE_SOURCE_DIR}" "utils" AMALGAMATION_UTILS)
   ADD_OSQUERY_LIBRARY_CORE(osquery_amalgamation ${AMALGAMATION_UTILS})
+endif()
+
+if(LINUX)
+  ADD_OSQUERY_LINK_CORE("libgcc_s")
+  ADD_OSQUERY_LINK_ADDITIONAL("uuid")
+  if(DEFINED ENV{OSQUERY_BUILD_SHARED})
+    ADD_OSQUERY_LINK_ADDITIONAL("-L${BUILD_DEPS}/legacy/lib")
+  endif()
+  if(NOT DEFINED ENV{OSQUERY_BUILD_LINK_SHARED})
+    ADD_OSQUERY_LINK_CORE("-static-libstdc++")
+    ADD_OSQUERY_LINK_ADDITIONAL("-static-libstdc++")
+  endif()
+  # For Ubuntu/CentOS packages add the build SHA1.
+  ADD_OSQUERY_LINK_CORE("-Wl,--build-id")
 endif()
 
 # Bubble the subdirectory (component) sources and links for this build.
@@ -198,7 +201,7 @@ set_target_properties(libosquery PROPERTIES COMPILE_FLAGS "${OSQUERY_LIBRARY_FLA
 # Create the dynamic libosquery.
 if(DEFINED ENV{OSQUERY_BUILD_SHARED})
   add_library(libosquery_shared SHARED main/lib.cpp ${OSQUERY_OBJECTS})
-  target_link_libraries(libosquery_shared ${OSQUERY_LIBS})
+  target_link_libraries(libosquery_shared PRIVATE ${OSQUERY_LIBS})
   set_target_properties(libosquery_shared PROPERTIES COMPILE_FLAGS "${OSQUERY_LIBRARY_FLAGS}")
   set_target_properties(libosquery_shared PROPERTIES OUTPUT_NAME osquery)
 endif()
@@ -236,13 +239,14 @@ if(NOT ${OSQUERY_BUILD_SDK_ONLY})
   # Create the dynamic libosquery_additional.
   if(DEFINED ENV{OSQUERY_BUILD_SHARED})
     add_library(libosquery_additional_shared SHARED ${OSQUERY_ADDITIONAL_SOURCES})
-    target_link_libraries(libosquery_additional_shared ${OSQUERY_ADDITIONAL_LINKS})
+    target_link_libraries(libosquery_additional_shared PRIVATE ${OSQUERY_ADDITIONAL_LINKS})
     set_target_properties(libosquery_additional_shared PROPERTIES OUTPUT_NAME osquery_additional)
   endif()
 
   add_executable(shell devtools/shell.cpp main/shell.cpp)
   ADD_DEFAULT_LINKS(shell TRUE)
   SET_OSQUERY_COMPILE(shell "${CXX_COMPILE_FLAGS}")
+  target_link_libraries(shell linenoise)
   set_target_properties(shell PROPERTIES OUTPUT_NAME osqueryi)
 
   add_executable(daemon main/${PROCESS_FAMILY}/daemon.cpp)
